move HPP response test into a seperate file within the HPP context
