+++ test/unit/hpp_test.rb
@@ -50,143 +50,6 @@ class HppTest < Minitest::Test
     )
   end
 
-  def test_autodetected_redirect_url
-    request = Adyen::HPP::Request.new(@payment_attributes)
-    assert_equal 'https://test.adyen.com/hpp/select.shtml', request.url
-
-    Adyen.configuration.stubs(:autodetect_environment).returns('live')
-    assert_equal 'https://live.adyen.com/hpp/select.shtml', request.url
-  end
-
-  def test_explicit_redirect_url
-    assert_equal 'https://test.adyen.com/hpp/select.shtml',
-      Adyen::HPP::Request.new(@payment_attributes, skin: :skin1, environment: :test).url
-    assert_equal 'https://live.adyen.com/hpp/select.shtml',
-      Adyen::HPP::Request.new(@payment_attributes, skin: :skin1, environment: :live).url
-    assert_equal 'https://test.adyen.com/hpp/select.shtml',
-      Adyen::HPP::Request.new(@payment_attributes, skin: :skin2, environment: :test).url
-    assert_equal 'https://live.adyen.com/hpp/select.shtml',
-      Adyen::HPP::Request.new(@payment_attributes, skin: :skin2, environment: :live).url
-    assert_equal 'https://test.adyen.com/hpp/select.shtml',
-      Adyen::HPP::Request.new(@payment_attributes, environment: :test).url
-    assert_equal 'https://live.adyen.com/hpp/select.shtml',
-      Adyen::HPP::Request.new(@payment_attributes, environment: :live).url
-  end
-
-  def test_redirect_url_for_different_payment_flows
-    request = Adyen::HPP::Request.new(@payment_attributes, environment: :test)
-
-    Adyen.configuration.payment_flow = :select
-    assert_equal 'https://test.adyen.com/hpp/select.shtml', request.url
-
-    Adyen.configuration.payment_flow = :pay
-    assert_equal 'https://test.adyen.com/hpp/pay.shtml', request.url
-
-    Adyen.configuration.payment_flow = :details
-    assert_equal 'https://test.adyen.com/hpp/details.shtml', request.url
-  end
-
-  def test_redirect_url_for_custom_domain
-    request = Adyen::HPP::Request.new(@payment_attributes, environment: :test)
-
-    Adyen.configuration.payment_flow_domain = "checkout.mydomain.com"
-    assert_equal 'https://checkout.mydomain.com/hpp/select.shtml', request.url
-  end
-
-  def test_redirect_url_generation
-    attributes = {
-      :currency_code => 'GBP', :payment_amount => 10000, :ship_before_date => Date.parse('2015-10-26'),
-      :merchant_reference => 'Internet Order 12345', :session_validity => Time.parse('2015-10-26 10:30')
-    }
-
-    request = Adyen::HPP::Request.new(attributes)
-
-    processed_attributes = {
-      'currencyCode' => 'GBP', 'paymentAmount' => '10000', 'shipBeforeDate' => '2015-10-26',
-      'merchantReference' => 'Internet Order 12345', 'sessionValidity' => '2015-10-26T10:30:00Z',
-      'merchantAccount' => 'TestMerchant', 'skinCode' => @skin_code1, 'merchantSig' => 'ewDgqa+m3rMO6MOZfQ0ugWdwsu+otvRVBVujqGfgvb8='
-    }
-
-    redirect_uri = URI(request.redirect_url)
-    assert_match %r[^#{request.url}], redirect_uri.to_s
-
-    params = CGI.parse(redirect_uri.query)
-    processed_attributes.each do |key, value|
-      assert_equal value, params[key].first
-    end
-  end
-
-  def test_redirect_url_generation_explicit_skin_code_and_shared_secret
-    attributes = {
-      :currency_code => 'GBP', :payment_amount => 10000, :ship_before_date => Date.parse('2015-10-26'),
-      :merchant_reference => 'Internet Order 12345', :session_validity => Time.parse('2015-10-26 10:30'),
-      :skin_code => @skin_code1
-    }
-
-    request = Adyen::HPP::Request.new(attributes, shared_secret: @shared_secret_skin1)
-
-    processed_attributes = {
-      'currencyCode' => 'GBP', 'paymentAmount' => '10000', 'shipBeforeDate' => '2015-10-26',
-      'merchantReference' => 'Internet Order 12345', 'sessionValidity' => '2015-10-26T10:30:00Z',
-      'merchantAccount' => 'TestMerchant', 'skinCode' => @skin_code1, 'merchantSig' => 'ewDgqa+m3rMO6MOZfQ0ugWdwsu+otvRVBVujqGfgvb8='
-    }
-
-    redirect_uri = URI(request.redirect_url)
-    assert_match %r[^#{request.url}], redirect_uri.to_s
-
-    params = CGI.parse(redirect_uri.query)
-    processed_attributes.each do |key, value|
-      assert_equal value, params[key].first
-    end
-  end
-
-  def test_redirect_url_generation_with_direct_skin_details
-    attributes = {
-      :currency_code => 'GBP', :payment_amount => 10000, :ship_before_date => Date.parse('2015-10-26'),
-      :merchant_reference => 'Internet Order 12345', :session_validity => Time.parse('2015-10-26 10:30'),
-      :merchant_account => 'OtherMerchant'
-    }
-
-    request = Adyen::HPP::Request.new(attributes, skin: :skin2)
-
-    processed_attributes = {
-      'currencyCode' => 'GBP', 'paymentAmount' => '10000', 'shipBeforeDate' => '2015-10-26',
-      'merchantReference' => 'Internet Order 12345', 'sessionValidity' => '2015-10-26T10:30:00Z',
-      'merchantAccount' => 'OtherMerchant', 'skinCode' => @skin_code2, 'merchantSig' => 'uS/tdqapxD8rQKBRzKQ9wOIiOFRmcOR3HsC8CO15Zto='
-    }
-
-    redirect_uri = URI(request.redirect_url)
-    assert_match %r[^#{request.url}], redirect_uri.to_s
-
-    params = CGI.parse(redirect_uri.query)
-    processed_attributes.each do |key, value|
-      assert_equal value, params[key].first
-    end
-  end
-
-  def test_payment_methods_url_generation
-    attributes = {
-      :currency_code => 'GBP', :payment_amount => 10000, :ship_before_date => Date.parse('2015-10-26'),
-      :merchant_reference => 'Internet Order 12345', :session_validity => Time.parse('2015-10-26 10:30')
-    }
-
-    request = Adyen::HPP::Request.new(attributes)
-
-    processed_attributes = {
-      'currencyCode' => 'GBP', 'paymentAmount' => '10000', 'shipBeforeDate' => '2015-10-26',
-      'merchantReference' => 'Internet Order 12345', 'sessionValidity' => '2015-10-26T10:30:00Z',
-      'merchantAccount' => 'TestMerchant', 'skinCode' => @skin_code1, 'merchantSig' => 'ewDgqa+m3rMO6MOZfQ0ugWdwsu+otvRVBVujqGfgvb8='
-    }
-
-    payment_methods_uri = URI(request.payment_methods_url)
-    assert_match %r[^#{request.url(:directory)}], payment_methods_uri.to_s
-
-    params = CGI.parse(payment_methods_uri.query)
-    processed_attributes.each do |key, value|
-      assert_equal value, params[key].first
-    end
-  end
-
   def test_has_valid_signature
     params = {
       'authResult' => 'AUTHORISED', 'pspReference' => '1211992213193029',
